Merge removed, Crafting added
I discarded the idea of merging cards, because I am not allowed/able to modify the damage of the cards. Instead you can now give away a card and receive an element fragment. With 2 element fragments you can craft a card of the same element type with a min dmg of 40.